Move the `default_authority` method to `h.security`
