+++ h/app.py
@@ -78,7 +78,6 @@ def includeme(config):
 
     # Core site modules
     config.include("h.assets")
-    config.include("h.auth")
     config.include("h.db")
     config.include("h.eventqueue")
     config.include("h.form")
